split into two files
sep of core functionality from gui
